+++ src/org/ojalgo/array/Primitive64Array.java
@@ -31,7 +31,7 @@ import java.util.stream.StreamSupport;
 
 import org.ojalgo.array.operation.AMAX;
 import org.ojalgo.array.operation.AXPY;
-import org.ojalgo.array.operation.Raw1D;
+import org.ojalgo.array.operation.COPY;
 import org.ojalgo.function.BinaryFunction;
 import org.ojalgo.function.BinaryFunction.FixedFirst;
 import org.ojalgo.function.BinaryFunction.FixedSecond;
@@ -450,7 +450,7 @@ public class Primitive64Array extends PrimitiveArray {
     }
 
     protected final double[] copyOfData() {
-        return Raw1D.copyOf(data);
+        return COPY.copyOf(data);
     }
 
     @Override