+++ src/org/ojalgo/array/operation/BLAS1.java
@@ -29,6 +29,17 @@ import org.ojalgo.scalar.Quaternion;
 import org.ojalgo.scalar.RationalNumber;
 
 /**
+ * <p>
+ * Contents in this package loosely corresponds to BLAS. The exact selection of operations and their API:s are
+ * entirely dictated by the requirements of the various {@linkplain org.ojalgo.matrix.store.MatrixStore}
+ * implementations.
+ * </p>
+ * <ul>
+ * <li>http://en.wikipedia.org/wiki/Basic_Linear_Algebra_Subprograms</li>
+ * <li>http://www.netlib.org/blas/</li>
+ * <li>http://www.netlib.org/blas/faq.html</li>
+ * <li>http://www.netlib.org/lapack/lug/node145.html</li>
+ * </ul>
  * Basic Linear Algebra Subprograms (BLAS) Level 1 contains vector operations.
  * <ul>
  * <li><a href="http://en.wikipedia.org/wiki/Basic_Linear_Algebra_Subprograms#Level_1">BLAS Level 1 @