+++ src/org/ojalgo/array/operation/BLAS3.java
@@ -22,6 +22,17 @@
 package org.ojalgo.array.operation;
 
 /**
+ * <p>
+ * Contents in this package loosely corresponds to BLAS. The exact selection of operations and their API:s are
+ * entirely dictated by the requirements of the various {@linkplain org.ojalgo.matrix.store.MatrixStore}
+ * implementations.
+ * </p>
+ * <ul>
+ * <li>http://en.wikipedia.org/wiki/Basic_Linear_Algebra_Subprograms</li>
+ * <li>http://www.netlib.org/blas/</li>
+ * <li>http://www.netlib.org/blas/faq.html</li>
+ * <li>http://www.netlib.org/lapack/lug/node145.html</li>
+ * </ul>
  * Basic Linear Algebra Subprograms (BLAS) Level 3 contains matrix-matrix operations.
  * <ul>
  * <li><a href="http://en.wikipedia.org/wiki/Basic_Linear_Algebra_Subprograms#Level_3">BLAS Level 3 @