+++ src/org/ojalgo/array/operation/ModifyAll.java
@@ -21,6 +21,8 @@
  */
 package org.ojalgo.array.operation;
 
+import java.util.function.DoubleUnaryOperator;
+
 public final class ModifyAll extends ArrayOperation {
 
     public static final ModifyAll SETUP = new ModifyAll();
@@ -30,6 +32,36 @@ public final class ModifyAll extends ArrayOperation {
     @Override
     public int threshold() {
         return THRESHOLD;
+    }
+
+    public static void modifyAll(final double[][] target, final DoubleUnaryOperator function) {
+        int tmpLength = target.length;
+        for (int i = 0; i < tmpLength; i++) {
+            int tmpInnerLength = target[i].length;
+            for (int j = 0; j < tmpInnerLength; j++) {
+                target[i][j] = function.applyAsDouble(target[i][j]);
+            }
+        }
+    }
+
+    public static void modifyColumn(final double[][] target, final int row, final int col, final DoubleUnaryOperator function) {
+        for (int i = row, limit = target.length; i < limit; i++) {
+            target[i][col] = function.applyAsDouble(target[i][col]);
+        }
+    }
+
+    public static void modifyDiagonal(final double[][] target, final int row, final int col, final DoubleUnaryOperator function) {
+        int tmpLength = target.length;
+        for (int ij = 0; ((row + ij) < tmpLength) && ((col + ij) < target[row + ij].length); ij++) {
+            target[row + ij][col + ij] = function.applyAsDouble(target[row + ij][col + ij]);
+        }
+    }
+
+    public static void modifyRow(final double[][] target, final int row, final int col, final DoubleUnaryOperator function) {
+        double[] targetRow = target[row];
+        for (int j = col, limit = targetRow.length; j < limit; j++) {
+            targetRow[j] = function.applyAsDouble(targetRow[j]);
+        }
     }
 
 }