+++ src/org/ojalgo/array/operation/package-info.java
@@ -1,15 +1,8 @@
 /**
  * <p>
- * Contents in this package loosely corresponds to BLAS. The exact selection of operations and their API:s are
- * entirely dictated by the requirements of the various {@linkplain org.ojalgo.matrix.store.MatrixStore}
- * implementations.
+ * Everythinng in this package should be considered library-private – it should only be used by other parts of
+ * ojAlgo!
  * </p>
- * <ul>
- * <li>http://en.wikipedia.org/wiki/Basic_Linear_Algebra_Subprograms</li>
- * <li>http://www.netlib.org/blas/</li>
- * <li>http://www.netlib.org/blas/faq.html</li>
- * <li>http://www.netlib.org/lapack/lug/node145.html</li>
- * </ul>
  *
  * @author apete
  */